+++ include/indigo/ui/UILoader.h
@@ -42,7 +42,7 @@ namespace Indigo
 
 		void unload();
 
-		OOBase::SharedPtr<UILayer> find_layer(const char* name, size_t len = size_t(-1)) const;
+		OOBase::SharedPtr<UILayer> find_layer(const char* name, size_t len = -1) const;
 
 		OOBase::SharedPtr<Window> window() const { return m_wnd; }
